+++ src/common/Clock.h
@@ -19,6 +19,18 @@
 
 #include <time.h>
 
-utime_t ceph_clock_now();
+static inline utime_t ceph_clock_now()
+{
+#if defined(__linux__)
+  struct timespec tp;
+  clock_gettime(CLOCK_REALTIME, &tp);
+  utime_t n(tp);
+#else
+  struct timeval tv;
+  gettimeofday(&tv, nullptr);
+  utime_t n(&tv);
+#endif
+  return n;
+}
 
 #endif
